requirements.txt
cffi==1.13.2
cryptography==2.8
Django==2.1.3
django-ckeditor==5.7.1
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-redis==4.10.0
django-simple-captcha==0.5.12
mysqlclient==1.4.6
Pillow==5.2.0
pycparser==2.19
PyMySQL==0.9.3
pytz==2019.2
redis==3.3.11
six==1.13.0
